notebooks/requirements.txt
antropy
fastapi
ipython
jupyterlab
matplotlib
more_itertools
multiprocess
nest-asyncio
numpy
openai
pandas
pycaret[analysis]
pycaret[mlops]
pycaret
pydantic
python-dotenv
scipy
seaborn
tqdm
uvicorn
